Allow multiple USB implementations.
